TEST: Initial mockup implementation of boost::asio::serial_port.
Extensive testing can now be done, but some changes needs to be performed
to allow opening a device from it's name with multiple
basic_serial_port<mockup_serial_port_service> and reading writing to it.

This can be done with a central storage of impl based on the device name.
This way we will totally mimic OSes behaviour.