BUGFIX: #5 thread datarace was a real problem, the error was that code which writes and reads from both ends of the serial port could result in a never-ending read loop, as the first handler would eat all bytes and not the second.
Therefore the current fix, is to read only from the io_services which
are different than the sender, so that the sender doesn't read his own
write before the receiver, which would then otherwise loop until something
else get written.

This is not perfect, we should as well duplicate the data for all readers
waiting. Will open an issue about it.