BUGFIX: Support clean cancellation and stop reading when port get closed.
In order to allow stopping the proponents attached to a port, when it gets
closed simply leave the read handler.

Make cancellation ephemeral to allow failed read because of timeout to read
again without need to reopen port.

This fixes unit-tests which were sometimes hanging or failing in our dependee
lib-cpp-slc.